WeightData {Simple}	R Documentation

Data on height, weight and age of children

Description

The WeightData data frame has 5337 rows and 5 columns.

Data set from NHANES data that contains weight, height and age information of children sampled from the United States.

Usage

data(WeightData)

Format

This data frame contains the following columns:

age

age in months of child

weight

weight in pounds

height

height in inches

gender

Male = 1, Female = 2

sample.weight

a weight to allow for adjustment to oversampling procedures

Details

This data set is excerpted from the NHANES III national survey. See the NHANES web site for much more information about the data.

Source

NHANES III national survey conducted 1999-2001, http://www.cdc.gov/nchs/nhanes.htm

References

Data is used to make growth charts.

Examples

data(WeightData)
age.yrs = floor(WeightData$age/12)
boxplot(WeightData$weight ~ age.yrs)

[Package Simple version 0.6 Index]