densityplot {Simple}	R Documentation

Plots densities of data

Description

Allows one to compare empirical densities of different distributions in a simple manner. The density is used as graphs with multiple histograms are too crowded. The usage is similar to side-by-side boxplots.

Usage

densityplot(x, ...)

Arguments

x 	 x may be a sequence of data vectors (eg. x,y,z), a data frame with numeric column vectors or a model formula

... 	 You can pass in a bandwidth argument such as bw="SJ". See density for details. A legend will be placed for you automatically. To overide the positioning set do.legend="manual". To skip the legend, set do.legend=FALSE.

Value

Makes a plot

Author(s)

John Verzani

References

Basically a modified boxplot function. As well it should be as it serves the same utility: comparing distributions.

See Also

boxplot,violinplot,density

Examples

## taken from boxplot
## using a formula
data(InsectSprays)
densityplot(count ~ spray, data = InsectSprays)
## on a matrix (data frame)
mat <- cbind(Uni05 = (1:100)/21, Norm = rnorm(100),
             T5 = rt(100, df = 5), Gam2 = rgamma(100, shape = 2))
densityplot(data.frame(mat))

[Package Simple version 0.6 Index]