homedata {Simple}	R Documentation

Maplewood NJ assessed values for years 1970 and 2000

Description

The homedata data frame has 6841 rows and 2 columns.

Data set containing assessed values of homes in Maplewood NJ for the years 1970 and 2000. The properties were not officially assessed during that time and it is interesting to see the change in percentage appreciation.

Usage

data(homedata)

Format

This data frame contains the following columns:

y1970

a numeric vector

y2000

a numeric vector

Source

Maplewood Reval

Examples

data(homedata)
plot(homedata)

[Package Simple version 0.6 Index]