homeprice {Simple}	R Documentation

Sale price of homes in New Jersey in the year 2001

Description

The homeprice data frame has 29 rows and 7 columns.

Usage

data(homeprice)

Format

This data frame contains the following columns:

list

list price of home (in thousands)

sale

actual sale price

full

Number of full bathrooms

half

number of half bathrooms

bedrooms

number of bedrooms

rooms

total number of rooms

neighborhood

Subjective assessment of neighborhood on scale of 1-5

Details

This dataset is a random sampling of the homes sold in Maplewood, NJ during the year 2001. Of course the prices will either seem incredibly high or fantastically cheap depending on where you live, and if you have recently purchased a home.

Source

Source Burgdorff Realty.

Examples

data(homeprice)
plot(homeprice$sale,homeprice$list)
abline(lm(homeprice$list~homeprice$sale))

[Package Simple version 0.6 Index]