simple.eda {Simple}	R Documentation

Simple function to plot histogram, boxplot and normal plot

Description

Simply plots histogram, boxplot and normal plot for experimental data analysis.

Usage

simple.eda(x)

Arguments

x 	 a vector of data

Value

Just does the plots. No return value

Author(s)

John Verzani

References

Inspired by S-Plus documentation

See Also

hist,boxplot,qnorm

Examples

  x<- rnorm(100,5,10)
  simple.eda(x)

[Package Simple version 0.6 Index]