simple.eda.ts {Simple}	R Documentation

Makes 3 useful graphs for eda of times series

Description

This makes 3 graphs to check for serial correlation in data. The graphs are a sequential plot (i vs X_i), a lag plot (plotting X_i vs X_{i+k} where k=1 by default) and an autocorrelation plot from the times series ("ts") package.

Usage

simple.eda.ts(x, lag=1)

Arguments

x 	 a univariate vector of data

lag 	 a lag to give to the lag plot

Value

Makes the graph with 1 row, 3 columns

Author(s)

John Verzani

References

see href{http://www.itl.nist.gov/div898/handbook/eda/section3/eda34.htm} for more information on these and other plots for exploratory data analysis.

Examples

## The function is currently defined as

## look for no correlation
x <- rnorm(100);simple.eda.ts(x)
## you will find correlation here
simple.eda.ts(cumsum(x))

[Package Simple version 0.6 Index]