simple.freqpoly {Simple}	R Documentation

Simply plot histogram and frequency polygon

Description

Simply plot histogram and frequency polygon. Students do not need to know how to add lines to a histogram, and how to extract values.

Usage

simple.freqpoly(x, ...)

Arguments

x 	 a vector of data

... 	 arguments passed onto histogram

Value

returns just the plot

Author(s)

John Verzani

See Also

hist,density

Examples

x <- rt(100,4)
simple.freqpoly(x)

[Package Simple version 0.6 Index]