simple.hist.and.boxplot {Simple}	R Documentation

A function to plot both a histogram and a boxplot

Description

Simple function to plot both histogram and boxplot to compare

Usage

simple.hist.and.boxplot(x, ...)

Arguments

x 	 vector of univariate data

... 	 Arguments passed to the hist function

Value

Just prints the two graphs

Author(s)

John Verzani

See Also

hist,boxplot,layout

Examples

x<-rnorm(100)
simple.hist.and.boxplot(x)

[Package Simple version 0.6 Index]