simple.scatterplot {Simple}	R Documentation

Simple scatter plot of x versus y with histograms of each

Description

Shows scatterplot of x vs y with histograms of each on sides of graph. As in the example from layout.

Usage

simple.scatterplot(x, y, ...)

Arguments

x 	 data vector

y 	 data vector

... 	 passed to plot command

Value

Returns the plot

Author(s)

John Verzani

See Also

layout

Examples

  x<-sort(rnorm(100))
  y<-sort(rt(100,3))
  simple.scatterplot(x,y)

[Package Simple version 0.6 Index]