simple.sim {Simple}	R Documentation

Simplify the process of simulation

Description

'simple.sim' is intended to make it a little easier to do simulations with R. Instead of writing a for loop, or dealing with column or row sums, a student can use this "simpler" interface.

Usage

simple.sim(no.samples, f, ...)

Arguments

no.samples 	 How many samples do you wish to generate

f 	 A function which generates a single random number from some distributions. simple.sim generates the rest.

... 	 parameters passed to f. It does not like named parameters.

Details

This is simply a wrapper for a for loop that uses the function f to create random numbers from some distribution.

Value

returns a vector of size no.samples

Author(s)

John Verzani

Examples

## First shows trivial (and very unnecessary usage)
## define a function f and then simulate
f<-function() rnorm(1)     # create a single random real number
sim <- simple.sim(100,f)   # create 100 random normal numbers
hist(sim)

## what does range look like?
f<- function (n,mu=0,sigma=1) {
  tmp <- rnorm(n,mu,sigma)
  max(tmp) - min(tmp)
}
sim <- simple.sim(100,f,5)
hist(sim)

[Package Simple version 0.6 Index]