simple.violinplot {Simple}	R Documentation

Plots violinplots instead of boxplots

Description

This function serves the same utility as side-by-side boxplots, only it provides more detail about the different distribution. It plots violinplots instead of boxplots. That is, instead of a box, it uses the density function to plot the density. For skewed distributions, the results look like "violins". Hence the name.

Usage

simple.violinplot(x, ...)

Arguments

x 	 Either a sequence of variable names, or a data frame, or a model formula

... 	 You can pass arguments to polygon with this. Notably, you can set the color to red with col='red', and a border color with border='blue'

Value

Returns a plot.

Author(s)

John Verzani

References

This is really the boxplot function from R/base with some minor adjustments

See Also

boxplot, simple.densityplot

Examples

## make a "violin"
x <- rnorm(100) ;x[101:150] <- rnorm(50,5)
simple.violinplot(x,col="brown")
f<-factor(rep(1:5,30))
## make a quintet. Note also choice of bandwidth
simple.violinplot(x~f,col="brown",bw="SJ")

[Package Simple version 0.6 Index]