skateranks {Simple}	R Documentation

Judges scores for disputed ice skating competition

Description

Judges scores from the disputed ice skating competition at the 2002 Winter olympics

Usage

data(skateranks)

Format

A data frame with 20 observations on the following 11 variables.

Name

a factor with levels Berankova/Diabola Berezhnaya/Sikharulidze Bestnadigova/Bestandif Chuvaeva/Palamarchuk Cobisi/DePra Ina/Zimmerman Kautz/Jeschke Krasitseva/Znachkov Langlois/Archetto Lariviere/Faustino Pang/Tong Petrova/Tikhonov Ponomareva/SWviridov Savchenko/Morozov Scott/Dulebohn Sele/Pelletier Shen/Zhao Totmianina/Marinin Zagorska/Siudek Zhang/Zhang

Country

a factor with levels Armenia Canada China Czech Germany Italy Poland Russia Slovakia US Ukraine Uzbekistan

Russia

a numeric vector

China

a numeric vector

US

a numeric vector

France

a numeric vector

Poland

a numeric vector

Canada

a numeric vector

Ukraine

a numeric vector

Germany

a numeric vector

Japan

a numeric vector

Examples

data(skateranks)

[Package Simple version 0.6 Index]